+++ src/staticMethods/queue.js
@@ -1,3 +1,5 @@
+import * as dom from '../utils/dom/index.js'
+
 // private global state for the queue feature
 let currentSteps = []
 
@@ -10,7 +12,6 @@ export const queue = function (steps) {
 
   const resetAndResolve = (resolve, value) => {
     currentSteps = []
-    document.body.removeAttribute('data-swal2-queue-step')
     resolve(value)
   }
 
@@ -37,7 +38,7 @@ export const queue = function (steps) {
 /*
  * Global function for getting the index of current popup in queue
  */
-export const getQueueStep = () => document.body.getAttribute('data-swal2-queue-step')
+export const getQueueStep = () => dom.getContainer().getAttribute('data-queue-step')
 
 /*
  * Global function for inserting a popup to the queue
