+++ test/qunit/params/progressSteps.js
@@ -1,9 +1,9 @@
-const { Swal } = require('../helpers')
+const { Swal, SwalWithoutAnimation } = require('../helpers')
 
 QUnit.test('first .swal2-progress-step is active by default', (assert) => {
   const done = assert.async()
 
-  Swal.queue([{
+  SwalWithoutAnimation.queue([{
     progressSteps: ['1', '2', '3'],
     onOpen: () => {
       assert.ok(Swal.getProgressSteps().querySelector('.swal2-progress-step').classList.contains('swal2-active-progress-step'))
